MAR-2068. Added tests to components except index.vue
