+++ tests/prismicSlices/pageParts/StartScreen/testsUtils.js
@@ -1,0 +1,7 @@
+export const getFakePropsByParams = params => ({
+  slice: {
+    primary: {
+      ...params,
+    },
+  },
+})
